convert local-storage to ts
